MYR-875: implemented PrimaryCoinMenu
